+++ Pandora/Paradigm/Structure/Ability/Morphable.hs
@@ -38,6 +38,8 @@ data Occurrence a = All a | First a
 
 data Vertical a = Up a | Down a
 
+data Horizontal a = Leftward a | Rightward a
+
 rotate :: forall mod struct . Morphable (Rotate mod) struct => struct ~> Morphing (Rotate mod) struct
 rotate = morphing . TT . Tag @(Rotate mod)